translate status code message on coffee module

+++ src/modules/coffees/useCases/deleteSpecification/DeleteSpecification.UseCase.ts
@@ -1,4 +1,5 @@
 import { inject, injectable } from "tsyringe";
+import { AppError } from "../../../../shared/errors/AppError";
 import { ISpecificationsRepository } from "../../repositories/interfaces/ISpecificationsRepository";
 
 @injectable()
@@ -9,6 +10,12 @@ class DeleteSpecificationUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const spec = await this.specificationsRepository.findById(id);
+
+    if (!spec) {
+      throw new AppError("Especificação não encontrada");
+    }
+
     await this.specificationsRepository.delete(id);
   }
 }
